app/src/main/java/com/mutkuensert/movee/data/movie/MovieApi.kt
package com.mutkuensert.movee.data.movie

import com.mutkuensert.movee.data.movie.model.MovieDetailsModel
import com.mutkuensert.movee.data.movie.model.MoviesNowPlayingModel
import com.mutkuensert.movee.data.movie.model.PopularMoviesModel
import com.mutkuensert.movee.data.movie.model.credits.MovieCredits
import com.mutkuensert.movee.util.API_KEY
import retrofit2.Response
import retrofit2.http.GET
import retrofit2.http.Path
import retrofit2.http.Query

interface MovieApi {

    @GET("movie/popular")
    suspend fun getPopularMovies(
        @Query("page") page: Int = 1,
        @Query("api_key") apiKey: String = API_KEY
    ): Response<PopularMoviesModel>

    @GET("movie/now_playing")
    suspend fun getMoviesNowPlaying(
        @Query("page") page: Int = 1,
        @Query("api_key") apiKey: String = API_KEY
    ): Response<MoviesNowPlayingModel>

    @GET("movie/{movie_id}")
    suspend fun getMovieDetails(
        @Path("movie_id") movieId: Int,
        @Query("api_key") apiKey: String = API_KEY
    ): Response<MovieDetailsModel>

    @GET("movie/{movie_id}/credits")
    suspend fun getMovieCredits(
        @Path("movie_id") movieId: Int,
        @Query("api_key") apiKey: String = API_KEY
    ): Response<MovieCredits>
}